requirements.txt
pdf2image
pytesseract
pillow
opencv-python-headless
sentence-transformers
faiss-cpu
google-generativeai
numpy
streamlit